requirements.txt
alembic==1.8.1
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
CoverAPI @ git+https://bitbucket.org/cover-webcie/coverapi.git@v0.1
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-CoverAPI @ git+https://bitbucket.org/cover-webcie/flask-coverapi.git@v1.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
pymdown-extensions==9.5
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.12
uWSGI==2.0.20
validators==0.20.0
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
